chore: break up helpers into builders and fetchers
