+++ tests/e2e/tracer/infrastructure.py
@@ -1,13 +1,13 @@
 from pathlib import Path
 
-from tests.e2e.utils import helpers
+from tests.e2e.utils.data_builder import build_service_name
 from tests.e2e.utils.infrastructure import BaseInfrastructureV2
 
 
 class TracerStack(BaseInfrastructureV2):
     # Maintenance: Tracer doesn't support dynamic service injection (tracer.py L310)
     # we could move after handler response or adopt env vars usage in e2e tests
-    SERVICE_NAME: str = helpers.build_service_name()
+    SERVICE_NAME: str = build_service_name()
 
     def __init__(self, handlers_dir: Path, feature_name: str = "tracer") -> None:
         super().__init__(feature_name, handlers_dir)
