+++ tests/e2e/utils/data_builder/traces.py
@@ -1,0 +1,2 @@
+def build_trace_default_query(function_name: str) -> str:
+    return f'service("{function_name}")'
