+++ tests/e2e/utils/data_fetcher/__init__.py
@@ -1,3 +1,5 @@
+from tests.e2e.utils.data_fetcher.logs import get_logs
+from tests.e2e.utils.data_fetcher.metrics import get_metrics
 from tests.e2e.utils.data_fetcher.traces import get_traces
 
-__all__ = ["get_traces"]
+__all__ = ["get_traces", "get_metrics", "get_logs"]
